Main Effect Importance

Extracts the weighted variances of the most relevant statistic (pd > pred_mean > y_mean > resid_mean > ale) from a "marginal" object. Serves as a simple measure of (main effect) importance. If partial dependence is used, the measure is almost identical to the importance measure proposed in the reference below.

main_effect_importance(x, by = NULL)

Arguments

x

Marginal object.

by

The statistic used to calculate the variance for. One of 'pd', 'pred_mean', 'y_mean', 'resid_mean', or 'ale'. By default (NULL), the first available of above list.

Details

If the importance cannot be calculated on a variable X (e.g., if X is constant), its importance is set to 0.

References

Greenwell, Brandon M., Bradley C. Boehmke, and Andrew J. McCarthy. 2018. A Simple and Effective Model-Based Variable Importance Measure. arXiv preprint. https://arxiv.org/abs/1805.04755.

See also

Examples

fit <- lm(Sepal.Length ~ ., data = iris)
xvars <- colnames(iris)[-1]
M <- marginal(fit, v = xvars, data = iris, y = "Sepal.Length", breaks = 5)
main_effect_importance(M)
#> Importance via weighted variance of 'pd'
#>  Sepal.Width Petal.Length  Petal.Width      Species 
#>   0.04161058   2.08704278   0.05597416   0.18456124